requirements.txt
discord.py==2.3.1
TikTokLive==6.5.2
Flask==2.3.2
firebase-admin==6.0.1
